+++ src/main/java/net/pneumono/locator_lodestones/INamed.java
@@ -1,0 +1,9 @@
+package net.pneumono.locator_lodestones;
+
+import java.util.Optional;
+import net.minecraft.text.Text;
+
+public interface INamed {
+    Optional<Text> getName();
+    void setName(Optional<Text> name);
+}